+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+#### Fixed
+
+- Call socket.destroy() when closing socket to fix intermitent condition where onerror is called when conn is closed by client.
+
 ## [2.1.1] - 2022-12-13
 
 ### Added