+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+#### Fixed
+
+- Call socket.destroy() when closing socket to fix intermitent condition where onerror is called when conn is closed by client.
+
 ### Changed
 
 - Breaking change: Removed support for end-of-life versions of Node.js. A minimum of Node.js 16 is now required.