+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Breaking change: Removed support for end-of-life versions of Node.js. A minimum of Node.js 16 is now required.
+- New overload for `AMQPWebSocketClient` constructor to allow setting optional parameters through an init object.
 
 ## [2.1.1] - 2022-12-13